updated weight function, new master function to run benchmark experiments, and matlab API in python to run Kun's code...

+++ labelshift.py
@@ -56,7 +56,7 @@ def estimate_target_dist(wt, ytrue_s,k):
 def beta_to_w(beta, y, k):
     w = []
     for i in range(k):
-        w.append(np.mean(beta[y == i]))
+        w.append(np.mean(beta[y.astype(int) == i]))
     w = np.array(w)
     return w
